user_django.py
Here's the transformed Django code for the same API function:

from django.http import JsonResponse
from django.views.decorators.csrf import csrf_exempt
from django.db import transaction
from .. import models, schemas, utils
from ..database import get_db

@csrf_exempt
@transaction.atomic
def create_user(request):
    if request.method == 'POST':
        user_data = schemas.UserCreate(**request.POST.dict())
        hashed_password = utils.hash(user_data.password)
        user_data.password = hashed_password
        
        exist = models.User.objects.filter(email=user_data.email).first()
        if exist is not None:
            return JsonResponse({'error': f"This email: {user_data.email} is already registered. Please use another email or renew your password for this one."}, status=409)
        
        exist = models.User.objects.filter(username=user_data.username).first()
        if exist is not None:
            return JsonResponse({'error': f"This username: {user_data.username} is already taken. Please choose another username."}, status=409)
        
        db_user = models.User(**user_data.dict())
        db_user.save()
        return JsonResponse(db_user.to_dict(), status=201)
    else:
        return JsonResponse({'error': 'Invalid Request Method'}, status=405)

Note: This code assumes that you have already defined the User model in your Django project and that you have imported the necessary modules.